backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
python-jose[cryptography]
passlib
python-multipart
gigachat
email-validator
requests
python-dotenv
google-api-python-client
google-auth
google-auth-httplib2
beautifulsoup4
lxml
